+++ cec2005real/mrandom.h
@@ -1,6 +1,6 @@
 #ifndef MRANDOM_H
 #define MRANDOM_H
 
-long double randomnormaldeviate();
+double randomnormaldeviate();
 
 #endif
